requirements.txt
# Venture
pyparsing>=1.5.7
flask>=0.10
requests>=1.2

# nose
nose-testconfig
nose-ignore-docstring
nose-cov

# VentureUnit
numpy
distribute>=0.7.3
matplotlib

# IPython Magics
ipython>=1.2
pyzmq

# lite
scipy>=0.13

# Plotting (optional)
patsy # Because ggplot needs this installed first ??
pandas
# ggplot
